Backend/requirements.txt
fastapi
uvicorn
pdf2docx
python-docx
reportlab
python-multipart
